refactor: 요구사항 테스트 패키지 구조 변경
