+++ src/test/java/roomescape/missionsteptest/MissionStep2Test.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.missionsteptest;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
+import roomescape.Reservation;
 
 import java.sql.Connection;
 import java.sql.SQLException;
